Abstrae funcionalidad de subida de fotos de gatites a una acción, refactorizada como Contexto.
