+++ src/App.js
@@ -6,15 +6,18 @@ import MainNavigation from 'navigation/Main'
 import { AuthProvider } from 'context/Auth'
 import { ThemeProvider } from 'context/Theme'
 import { DatabaseProvider } from 'context/Database'
+import { AccionesProvider } from 'context/Acciones'
 
 const App = () => {
   return (
     <AuthProvider>
       <DatabaseProvider>
-        <ThemeProvider>
-          { /* El contenedor de navegación que gestiona las diferentes pantallas. */ }
-          <MainNavigation />
-        </ThemeProvider>
+        <AccionesProvider>
+          <ThemeProvider>
+            {/* El contenedor de navegación que gestiona las diferentes pantallas. */}
+            <MainNavigation />
+          </ThemeProvider>
+        </AccionesProvider>
       </DatabaseProvider>
     </AuthProvider>
   )
