--- src/acciones/index.js
@@ -1,4 +1,0 @@
-export const subirFotosDeGatites = (nombres, archivos) => {
-  console.log(nombres)
-  console.log(archivos)
-}
