--- src/acciones/package.json
@@ -1,3 +1,0 @@
-{
-  "name": "acciones"
-}
